# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface StudentLoginRequest {
  schoolSlug: string;
  rollNumber: string;
  password: string;
}

export type AuthResponseUser = {
  id: number;
  name: string;
  username: string;
};

export interface AuthResponse {
  token: string;
  role: string;
  user: AuthResponseUser;
}

export interface MeResponse {
  id: number;
  name: string;
  role: string;
  schoolId?: number | null;
  schoolSlug?: string | null;
}

export interface School {
  id: number;
  name: string;
  slug: string;
  username: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateSchoolRequest {
  name: string;
  slug: string;
  username: string;
  password: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
}

export interface UpdateSchoolRequest {
  name?: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  isActive?: boolean;
  password?: string | null;
}

export interface SchoolProfile {
  id: number;
  name: string;
  slug: string;
  tagline?: string | null;
  about?: string | null;
  mission?: string | null;
  vision?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  logoUrl?: string | null;
  heroImageUrl?: string | null;
  yearsOfExperience?: number | null;
  principalMessage?: string | null;
  founderMessage?: string | null;
  presidentMessage?: string | null;
  feeStructure?: string | null;
  facilities?: string | null;
  mapUrl?: string | null;
  socialFacebook?: string | null;
  socialTwitter?: string | null;
  socialYoutube?: string | null;
}

export interface UpdateSchoolProfileRequest {
  tagline?: string | null;
  about?: string | null;
  mission?: string | null;
  vision?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  logoUrl?: string | null;
  heroImageUrl?: string | null;
  yearsOfExperience?: number | null;
  principalMessage?: string | null;
  founderMessage?: string | null;
  presidentMessage?: string | null;
  feeStructure?: string | null;
  facilities?: string | null;
  mapUrl?: string | null;
  socialFacebook?: string | null;
  socialTwitter?: string | null;
  socialYoutube?: string | null;
}

export interface Notification {
  id: number;
  title: string;
  content: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateNotificationRequest {
  title: string;
  content: string;
}

export interface Student {
  id: number;
  name: string;
  rollNumber: string;
  className?: string | null;
  section?: string | null;
  fatherName?: string | null;
  motherName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  enrollmentDate?: string | null;
  isActive: boolean;
  hasChangedPassword: boolean;
}

export interface StudentProfile {
  id: number;
  name: string;
  rollNumber: string;
  className?: string | null;
  section?: string | null;
  fatherName?: string | null;
  motherName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  enrollmentDate?: string | null;
  schoolName: string;
  schoolSlug: string;
  hasChangedPassword: boolean;
}

export interface CreateStudentRequest {
  name: string;
  rollNumber: string;
  className?: string | null;
  section?: string | null;
  fatherName?: string | null;
  motherName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  enrollmentDate?: string | null;
}

export interface UpdateStudentRequest {
  name?: string;
  className?: string | null;
  section?: string | null;
  fatherName?: string | null;
  motherName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  enrollmentDate?: string | null;
  isActive?: boolean;
}

export interface Result {
  id: number;
  studentId: number;
  studentName?: string | null;
  rollNumber?: string | null;
  className?: string | null;
  subject: string;
  marks: number;
  maxMarks: number;
  grade?: string | null;
  examType?: string | null;
  examDate?: string | null;
  remarks?: string | null;
}

export interface UploadResultResponse {
  success: boolean;
  inserted: number;
  errors: string[];
}

export interface GalleryImage {
  id: number;
  url: string;
  caption?: string | null;
  createdAt: string;
}

export interface AddGalleryImageRequest {
  url: string;
  caption?: string | null;
}

export interface PublicSchool {
  id: number;
  name: string;
  slug: string;
  tagline?: string | null;
  city?: string | null;
  state?: string | null;
  logoUrl?: string | null;
}

export interface PublicSchoolDetail {
  id: number;
  name: string;
  slug: string;
  tagline?: string | null;
  about?: string | null;
  mission?: string | null;
  vision?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  logoUrl?: string | null;
  heroImageUrl?: string | null;
  yearsOfExperience?: number | null;
  principalMessage?: string | null;
  founderMessage?: string | null;
  presidentMessage?: string | null;
  feeStructure?: string | null;
  facilities?: string | null;
  mapUrl?: string | null;
  socialFacebook?: string | null;
  socialTwitter?: string | null;
  socialYoutube?: string | null;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export type UploadResultsBody = {
  file?: Blob;
};