# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/superadmin/login:
    post:
      operationId: superadminLogin
      tags: [auth]
      summary: Super admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/school/login:
    post:
      operationId: schoolLogin
      tags: [auth]
      summary: School admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/student/login:
    post:
      operationId: studentLogin
      tags: [auth]
      summary: Student login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentLoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
...
  /public/schools/{slug}:
    get:
      operationId: getPublicSchool
      tags: [public]
      summary: Get public school by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public school profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSchoolDetail"
        "404":
          description: School not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /public/schools/{slug}/notifications:
    get:
      operationId: getPublicSchoolNotifications
      tags: [public]
      summary: Get public notifications for a school
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /public/schools/{slug}/gallery:
    get:
      operationId: getPublicSchoolGallery
      tags: [public]
      summary: Get public gallery for a school
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"

  # ---- STUDENT ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    StudentLoginRequest:
      type: object
      properties:
        schoolSlug:
          type: string
        rollNumber:
          type: string
        password:
          type: string
      required:
        - schoolSlug
        - rollNumber
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            username:
              type: string
          required:
            - id
            - name
            - username
      required:
        - token
        - role
        - user
    MeResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        schoolId:
          type: integer
          nullable: true
        schoolSlug:
          type: string
          nullable: true
      required:
        - id
        - name
        - role
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - isActive
        - createdAt
...
    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - url
        - createdAt
...
    PublicSchoolDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
          nullable: true
        about:
          type: string
          nullable: true
        mission:
          type: string
          nullable: true
        vision:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        yearsOfExperience:
          type: integer
          nullable: true
        principalMessage:
          type: string
          nullable: true
        founderMessage:
          type: string
          nullable: true
        presidentMessage:
          type: string
          nullable: true
        feeStructure:
          type: string
          nullable: true
        facilities:
          type: string
          nullable: true
        mapUrl:
          type: string
          nullable: true
        socialFacebook:
          type: string
          nullable: true
        socialTwitter:
          type: string
          nullable: true
        socialYoutube:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
...