# lib/api-zod/src/generated/types/meResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MeResponse {
  id: number;
  name: string;
  role: string;
  schoolId?: number | null;
  schoolSlug?: string | null;
}