# lib/api-zod/src/generated/types/studentLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentLoginRequest {
  schoolSlug: string;
  rollNumber: string;
  password: string;
}